refactor(solver): split generic application helpers
