+++ crates/tsz-solver/src/relations/subtype/rules/generics.rs
@@ -10,31 +10,20 @@
 use super::super::{SubtypeChecker, SubtypeResult, TypeResolver};
 pub(crate) use super::mapped_chain::flatten_mapped_chain;
 use crate::def::DefId;
-use crate::diagnostics::SubtypeFailureReason;
 use crate::instantiation::instantiate::fill_application_defaults;
 use crate::types::{MappedModifier, MappedType, TypeData, TypeParamInfo};
-use crate::types::{MappedTypeId, SymbolRef, TypeApplicationId, TypeId, Variance};
+use crate::types::{MappedTypeId, SymbolRef, TypeApplicationId, TypeId};
 use crate::visitor::{
     application_id, array_element_type, contains_type_parameter_named, index_access_parts,
     intersection_list_id, is_empty_object_type, keyof_inner_type, mapped_type_id, object_shape_id,
     object_with_index_shape_id, tuple_list_id, type_param_info, union_list_id,
 };
 use crate::visitors::visitor_predicates::is_primitive_type;
-use std::sync::Arc;
 
-/// Maximum nesting depth, per generic `DefId`, for the one-sided application
-/// expansion relation paths (`App <: T` and `T <: App`).
-///
-/// This mirrors tsc's `isDeeplyNestedType` recursion-identity bailout (default
-/// `maxDepth = 5`). When the same generic definition is re-introduced by its own
-/// structural expansion past this depth, the relation is assumed related
-/// (`Ternary.Maybe`) instead of expanding further. The two-sided `App <: App`
-/// comparison already bounds recursion through `def_guard`; this constant bounds
-/// the previously-unguarded one-sided paths so a generic alias whose expansion
-/// keeps re-introducing itself (recursive mapped/conditional/template
-/// compositions) terminates cheaply rather than driving a terminating-but-
-/// exponential evaluate<->subtype expansion.
-pub(crate) const ONE_SIDED_APP_EXPANSION_MAX_DEPTH: u32 = 5;
+#[path = "generics_application_helpers.rs"]
+mod generics_application_helpers;
+#[cfg(test)]
+pub(crate) use generics_application_helpers::ONE_SIDED_APP_EXPANSION_MAX_DEPTH;
 
 fn args_contain_type_parameters(
     interner: &dyn crate::construction::TypeDatabase,
@@ -45,275 +34,6 @@ fn args_contain_type_parameters(
 }
 
 impl<'a, R: TypeResolver> SubtypeChecker<'a, R> {
-    /// Enter a one-sided application expansion for `def_id`.
-    ///
-    /// Returns `true` if expansion may proceed and `false` when the same
-    /// generic definition is already nested at or beyond
-    /// [`ONE_SIDED_APP_EXPANSION_MAX_DEPTH`] (the caller should bail to
-    /// [`Self::depth_result`]). Callers that receive `true` must pair this with
-    /// [`Self::leave_app_expansion_depth`] once the expansion completes.
-    pub(crate) fn enter_app_expansion_depth(&mut self, def_id: DefId) -> bool {
-        let depth = self.app_expand_depth.get(&def_id).copied().unwrap_or(0);
-        if depth >= ONE_SIDED_APP_EXPANSION_MAX_DEPTH {
-            return false;
-        }
-        self.app_expand_depth.insert(def_id, depth + 1);
-        true
-    }
-
-    /// Leave a one-sided application expansion previously entered via
-    /// [`Self::enter_app_expansion_depth`].
-    pub(crate) fn leave_app_expansion_depth(&mut self, def_id: DefId) {
-        if let Some(depth) = self.app_expand_depth.get_mut(&def_id) {
-            *depth = depth.saturating_sub(1);
-        }
-    }
-
-    fn iterator_protocol_mismatch_for_same_application_family(
-        &mut self,
-        source_type: TypeId,
-        target_type: TypeId,
-    ) -> bool {
-        let Some(query_db) = self.query_db else {
-            return false;
-        };
-
-        let iterator_mismatch = |checker: &mut Self, is_async: bool| {
-            let source_eval = checker.evaluate_type(source_type);
-            let target_eval = checker.evaluate_type(target_type);
-            let source_info = crate::operations::get_iterator_info(query_db, source_eval, is_async)
-                .or_else(|| crate::operations::get_iterator_info(query_db, source_type, is_async));
-            let target_info = crate::operations::get_iterator_info(query_db, target_eval, is_async)
-                .or_else(|| crate::operations::get_iterator_info(query_db, target_type, is_async));
-            source_info
-                .zip(target_info)
-                .is_some_and(|(source, target)| {
-                    !checker
-                        .check_subtype(source.yield_type, target.yield_type)
-                        .is_true()
-                        || !checker
-                            .check_subtype(source.return_type, target.return_type)
-                            .is_true()
-                        || !checker
-                            .check_subtype(target.next_type, source.next_type)
-                            .is_true()
-                })
-        };
-
-        iterator_mismatch(self, false) || iterator_mismatch(self, true)
-    }
-
-    fn application_cycle_with_concrete_differing_args_is_unsound(
-        &self,
-        s_app: &crate::types::TypeApplication,
-        t_app: &crate::types::TypeApplication,
-    ) -> bool {
-        if s_app.args == t_app.args {
-            return false;
-        }
-
-        if self.application_type_args_are_unwitnessed(s_app)
-            && self.application_type_args_are_unwitnessed(t_app)
-        {
-            return false;
-        }
-
-        s_app.args.iter().chain(t_app.args.iter()).all(|&arg| {
-            !crate::contains_type_parameters(self.interner, arg)
-                && !crate::contains_this_type(self.interner, arg)
-        })
-    }
-
-    fn application_type_args_are_unwitnessed(&self, app: &crate::types::TypeApplication) -> bool {
-        let Some(def_id) = self.application_base_def_id(app.base) else {
-            return false;
-        };
-
-        use crate::caches::db::QueryDatabase;
-        let variances = self
-            .resolver
-            .get_type_param_variance(def_id)
-            .or_else(|| {
-                self.query_db
-                    .and_then(|db| QueryDatabase::get_type_param_variance(db, def_id))
-            })
-            .or_else(|| {
-                crate::relations::variance::compute_type_param_variances_with_resolver(
-                    self.interner,
-                    self.resolver,
-                    def_id,
-                )
-            });
-
-        variances.as_ref().is_some_and(|variances| {
-            variances.len() == app.args.len() && variances.iter().all(|v| v.is_independent())
-        })
-    }
-
-    fn application_base_def_id(&self, base: TypeId) -> Option<DefId> {
-        if base.is_intrinsic() {
-            return None;
-        }
-        match self.interner.lookup(base) {
-            Some(TypeData::Lazy(def_id)) => Some(def_id),
-            Some(TypeData::TypeQuery(sym_ref)) => {
-                let def_id = self.resolver.symbol_to_def_id(sym_ref)?;
-                matches!(
-                    self.resolver.get_def_kind(def_id),
-                    Some(crate::def::DefKind::Interface | crate::def::DefKind::TypeAlias)
-                )
-                .then_some(def_id)
-            }
-            _ => None,
-        }
-    }
-
-    /// Resolve the per-type-parameter variance mask for a generic definition,
-    /// preferring declared/cached variances and computing (and caching) them
-    /// only when missing. Shared by the variance-aware relation fast path and
-    /// the same-generic error-elaboration path so both observe identical
-    /// variance facts.
-    pub(crate) fn resolve_application_variances(&self, def_id: DefId) -> Option<Arc<[Variance]>> {
-        use crate::caches::db::QueryDatabase;
-        self.resolver
-            .get_type_param_variance(def_id)
-            .or_else(|| {
-                self.query_db
-                    .and_then(|db| QueryDatabase::get_type_param_variance(db, def_id))
-            })
-            .or_else(|| {
-                let computed =
-                    crate::relations::variance::compute_type_param_variances_with_resolver(
-                        self.interner,
-                        self.resolver,
-                        def_id,
-                    );
-                if let (Some(db), Some(variances)) = (self.query_db, computed.as_ref()) {
-                    db.insert_type_param_variance(def_id, variances.clone());
-                }
-                computed
-            })
-    }
-
-    /// Explain a same-generic application failure (`C<A..>` vs `C<B..>`) by
-    /// comparing the differing type **arguments** directly, mirroring tsc.
-    ///
-    /// Structural rule: when source and target are applications of the same
-    /// generic target whose variances are reliably measured (no mapped-modifier
-    /// structural fallback, no unreliable rejection) and whose arguments are
-    /// concrete (no embedded type parameters), tsc reports the first
-    /// variance-failing type argument as a direct nested line — e.g.
-    /// `Type 'number' is not assignable to type 'string'.` — without the
-    /// `Types of property 'x' are incompatible.` wrapper that structural
-    /// expansion would produce.
-    ///
-    /// Returns `None` (so the caller falls back to structural elaboration)
-    /// whenever the relation itself would fall through to structural
-    /// comparison, keeping the existing property-based elaboration for those
-    /// shapes. This mirrors the conclusive-rejection conditions in
-    /// [`Self::check_application_to_application_subtype`].
-    pub(crate) fn explain_same_generic_type_arguments(
-        &mut self,
-        source: TypeId,
-        target: TypeId,
-    ) -> Option<SubtypeFailureReason> {
-        // Callers may pass raw (lazy) references or already-resolved types;
-        // resolve lazy aliases so a `Lazy(DefId)` wrapping an application is
-        // recognised, while leaving direct applications untouched.
-        let resolved_source = self.resolve_lazy_type(source);
-        let resolved_target = self.resolve_lazy_type(target);
-        let s_app_id = application_id(self.interner, resolved_source)?;
-        let t_app_id = application_id(self.interner, resolved_target)?;
-        let s_app = self.interner.type_application(s_app_id);
-        let t_app = self.interner.type_application(t_app_id);
-
-        if s_app.base != t_app.base || s_app.args.len() != t_app.args.len() {
-            return None;
-        }
-
-        // Type-parameter arguments make a variance rejection inconclusive: the
-        // expanded structural forms can introduce implicit index signatures
-        // (homomorphic mapped types) that change the outcome. The relation
-        // falls through to structural comparison for these, so the explanation
-        // must do the same.
-        if args_contain_type_parameters(self.interner, &s_app.args) {
-            return None;
-        }
-
-        let def_id = self.application_base_def_id(s_app.base)?;
-        let variances = self.resolve_application_variances(def_id)?;
-        if variances.len() != s_app.args.len() {
-            return None;
-        }
-
-        // Mapped-modifier (`needs_structural_fallback`) and indexed-access /
-        // intersection-normalization (`rejection_unreliable`) variances make a
-        // variance-based rejection unreliable; tsc reports those structurally
-        // (its variance carries the structural-fallback marker), so keep the
-        // property-based elaboration for them.
-        if variances
-            .iter()
-            .any(|v| v.needs_structural_fallback() || v.rejection_unreliable())
-        {
-            return None;
-        }
-
-        // `s_app`/`t_app` are owned `Arc`s, independent of `self`, so the
-        // arguments can be indexed directly across the `&mut self` relation
-        // calls below without cloning the argument vectors.
-        for (i, variance) in variances.iter().enumerate() {
-            let s_arg = s_app.args[i];
-            let t_arg = t_app.args[i];
-
-            // Orient the failing relation by the parameter's variance, matching
-            // the per-argument direction used by the relation fast path.
-            let failing_pair = if variance.is_invariant() {
-                if !self.check_subtype(s_arg, t_arg).is_true() {
-                    Some((s_arg, t_arg))
-                } else if !self.check_subtype(t_arg, s_arg).is_true() {
-                    Some((t_arg, s_arg))
-                } else {
-                    None
-                }
-            } else if variance.is_covariant() {
-                (!self.check_subtype(s_arg, t_arg).is_true()).then_some((s_arg, t_arg))
-            } else if variance.is_contravariant() {
-                (!self.check_subtype(t_arg, s_arg).is_true()).then_some((t_arg, s_arg))
-            } else {
-                // Independent: argument does not constrain the relation.
-                None
-            };
-
-            if let Some((fail_src, fail_tgt)) = failing_pair {
-                // The type-argument elaboration is only reliable when this
-                // parameter's variance comes from a *direct* usage (a property,
-                // function parameter, or return type). Variances synthesized
-                // purely from mapped-type / indexed-access positions lack
-                // `DIRECT_USAGE`: there the differing arguments can normalize to
-                // structurally distinct shapes, and tsc reports the structural
-                // member difference (e.g. a missing property) rather than the
-                // raw argument relation. Fall back to structural elaboration for
-                // those, matching tsc.
-                if !variance.has_direct_usage() {
-                    return None;
-                }
-                let nested = self.explain_failure(fail_src, fail_tgt).unwrap_or(
-                    SubtypeFailureReason::TypeMismatch {
-                        source_type: fail_src,
-                        target_type: fail_tgt,
-                    },
-                );
-                return Some(SubtypeFailureReason::TypeArgumentMismatch {
-                    source_arg: fail_src,
-                    target_arg: fail_tgt,
-                    nested_reason: Box::new(nested),
-                });
-            }
-        }
-
-        None
-    }
-
     /// Helper for resolving two Ref/TypeQuery symbols and checking subtype.
     ///
     /// Handles the common pattern of:
